lseek fixed

+++ kern/include/syscall.h
@@ -76,7 +76,7 @@ int sys_open(userptr_t filename, int flags, int *retval);
 int sys_write(int fd, userptr_t buf, size_t buflen, int *retval);
 int sys_read(int fd, userptr_t buf, size_t size, int *retval);
 int sys_close(int fd);
-int sys_lseek(int fd, off_t pos, int whence, int *retval);
+int sys_lseek(int fd, off_t pos, int whence, int64_t *retval);
 int sys_dup2(int oldfd, int newfd, int *retval);
